mcp-flow/flow-mcp-requirements.txt
# Flow MCP Server Requirements - Model Context Protocol Server

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0

# ChromaDB client for connecting to chroma server
chromadb>=0.4.22

# MCP (Model Context Protocol) dependencies
mcp>=1.0.0

# Web server for HTTP bridge
fastapi>=0.108.0
uvicorn[standard]>=0.24.0

# Async support
aiohttp>=3.9.1

# Logging
structlog>=23.2.0
